pyproject.toml
[project]
name = "authserver-py"
version = "0.1.0"
description = "Python Authentication & Authorization Server"
authors = [
    {name = "Caleb Beery"}
]
# readme = "README.md"  # Commented out since we don't have a README yet
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "alembic (>=1.16.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pyotp (>=2.9.0,<3.0.0)",
    "authlib (>=1.6.1,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "requests (>=2.32.4,<3.0.0)",
    "redis (>=6.4.0,<7.0.0)",
    "email-validator (>=2.2.0,<3.0.0)"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.1.0"
pytest-cov = "^6.2.1"
pytest-mock = "^3.14.1"
black = "^25.1.0"
flake8 = "^7.3.0"
isort = "^6.0.1"
mypy = "^1.17.1"
pre-commit = "^4.3.0"
ipython = "^9.4.0"
ipdb = "^0.13.13"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.16"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
